pyproject.toml
[tool.poetry]
name = "autotm"
version = "0.2.2"
description = "Automatic hyperparameters tuning for topic models (ARTM approach) using evolutionary algorithms"
authors = [
    "Khodorchenko Maria <mariyaxod@yandex.ru>",
    "Nikolay Butakov alipoov.nb@gmail.com"
]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://autotm.readthedocs.io/en/latest/"
repository = "https://github.com/ngc436/AutoTM"
packages = [
    {include = "autotm"}
]


[tool.poetry.dependencies]
python = ">=3.8, <3.10"
bigartm = "0.9.2"
protobuf = "<=3.20.0"
tqdm = "4.50.2"
numpy = "*"
PyYAML = "5.3.1"
dataclasses-json = "*"
mlflow = "*"
click = "8.0.1"
scipy = "*"
hyperopt = "*"
pymystem3 = "*"
nltk = "*"
plotly = "*"
spacy = ">=3.5"
spacy-langdetect = "*"
gensim = "4.1.2"
pandas = "*"
billiard = "*"
dill = "*"
pytest = "*"
celery = "4.4.7"
redis = "3.5.3"
jinja2 = "3.0"
PyMySQL = "*"
psycopg2-binary = "*"
pymongo = "3.11.3"
scikit-learn = "1.1.1"
#pydantic = "1.10.8"
pydantic = "2.6.0"

[tool.poetry.dev-dependencies]
black = "*"
sphinx = "*"
flake8 = "*"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.poetry.scripts]
autotmctl = 'autotm.main:cli'
fitness-worker = 'autotm.main_fitness_worker:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
